fix/Теперь форма полностью управляет временными значениями
