+++ src/components/article-params-form/ArticleParamsForm.tsx
@@ -5,7 +5,7 @@ import clsx from 'clsx';
 import styles from './ArticleParamsForm.module.scss';
 
 import { useOutsideClickClose } from './hooks/useOutsideClickClose';
-import React, { useRef } from 'react';
+import React, { useRef, useState } from 'react';
 
 import { FontSelector } from '../font-selector/FontSelector';
 import { FontSizeRadioButton } from '../font-size-radio-button/FontSizeRadioButton';
@@ -13,23 +13,18 @@ import { FontColorSelector } from '../font-color/FontColorSelector';
 import { Separator } from 'src/ui/separator';
 import { BackgroundColorSelector } from '../backgroundcolor-selector/BackgroundColorSelector';
 import { ContentWidthSelector } from '../content-width-selector/ContentWidthSelector';
-import { OptionType } from 'src/constants/articleProps';
+
+import {
+	ArticleStateType,
+	defaultArticleState,
+} from 'src/constants/articleProps';
 
 export type TArticleParamsForm = {
 	isOpen: boolean;
 	onOpenChange: () => void;
-	onApply: () => void;
+	onApply: (newState: ArticleStateType) => void;
 	onReset: () => void;
-	onFontChange: (option: OptionType) => void;
-	onSizeChange: (option: OptionType) => void;
-	onColorChange: (option: OptionType) => void;
-	onBackgroundColorChange: (option: OptionType) => void;
-	onWidthChange: (option: OptionType) => void;
-	selectedFont: OptionType;
-	selectedSize: OptionType;
-	selectedColor: OptionType;
-	selectedBackgroundColor: OptionType;
-	selectedWidth: OptionType;
+	currentState: ArticleStateType;
 };
 
 export const ArticleParamsForm = ({
@@ -37,28 +32,51 @@ export const ArticleParamsForm = ({
 	onOpenChange,
 	onApply,
 	onReset,
-	onFontChange,
-	onSizeChange,
-	onColorChange,
-	onBackgroundColorChange,
-	onWidthChange,
-	selectedFont,
-	selectedSize,
-	selectedColor,
-	selectedBackgroundColor,
-	selectedWidth,
+	currentState,
 }: TArticleParamsForm) => {
 	const formRef = useRef<HTMLElement>(null);
+
+	const [pendingState, setPendingState] =
+		useState<ArticleStateType>(currentState);
 
 	useOutsideClickClose({ ref: formRef, isOpen: isOpen, onClose: onOpenChange });
 
 	const handleSubmit = (e: React.FormEvent) => {
 		e.preventDefault();
-		onApply();
+		onApply(pendingState);
 	};
 	const handleReset = (e: React.FormEvent) => {
 		e.preventDefault();
+		setPendingState(defaultArticleState);
 		onReset();
+	};
+
+	const handleFontChange = (
+		fontFamilyOption: ArticleStateType['fontFamilyOption']
+	) => {
+		setPendingState((prev) => ({ ...prev, fontFamilyOption }));
+	};
+
+	const handleSizeChange = (
+		fontSizeOption: ArticleStateType['fontSizeOption']
+	) => {
+		setPendingState((prev) => ({ ...prev, fontSizeOption }));
+	};
+
+	const handleColorChange = (fontColor: ArticleStateType['fontColor']) => {
+		setPendingState((prev) => ({ ...prev, fontColor }));
+	};
+
+	const handleBackgroundColorChange = (
+		backgroundColor: ArticleStateType['backgroundColor']
+	) => {
+		setPendingState((prev) => ({ ...prev, backgroundColor }));
+	};
+
+	const handleWidthChange = (
+		contentWidth: ArticleStateType['contentWidth']
+	) => {
+		setPendingState((prev) => ({ ...prev, contentWidth }));
 	};
 
 	return (
@@ -73,20 +91,20 @@ export const ArticleParamsForm = ({
 					onReset={handleReset}>
 					<div className={styles.spacing_50}>
 						<FontSelector
-							selectedFont={selectedFont}
-							onFontChange={onFontChange}
+							selectedFont={pendingState.fontFamilyOption}
+							onFontChange={handleFontChange}
 						/>
 					</div>
 					<div className={styles.spacing_50}>
 						<FontSizeRadioButton
-							selectedSize={selectedSize}
-							onSizeChange={onSizeChange}
+							selectedSize={pendingState.fontSizeOption}
+							onSizeChange={handleSizeChange}
 						/>
 					</div>
 					<div className={styles.spacing_50}>
 						<FontColorSelector
-							selectedColor={selectedColor}
-							onColorChange={onColorChange}
+							selectedColor={pendingState.fontColor}
+							onColorChange={handleColorChange}
 						/>
 					</div>
 					<div className={styles.spacing_50}>
@@ -94,14 +112,14 @@ export const ArticleParamsForm = ({
 					</div>
 					<div className={styles.spacing_50}>
 						<BackgroundColorSelector
-							selectedBackgroundColor={selectedBackgroundColor}
-							onBackgroundColorChange={onBackgroundColorChange}
+							selectedBackgroundColor={pendingState.backgroundColor}
+							onBackgroundColorChange={handleBackgroundColorChange}
 						/>
 					</div>
 					<div className={styles.spacing_50}>
 						<ContentWidthSelector
-							selectedWidth={selectedWidth}
-							onWidthChange={onWidthChange}
+							selectedWidth={pendingState.contentWidth}
+							onWidthChange={handleWidthChange}
 						/>
 					</div>
 					<div className={styles.bottomContainer}>
